spec: Rebuild with openssl-compat-10
In XCP-ng we keep the openssl support because it used by:
- in unbound (native with openssl-1 support)
  - used by libreswan
    - used by openvswitch-ipsec (an optional feature enabled by XO)

For the records building libevent-2.0.21-stable along openssl-3.0.9 is
failing on:

    bufferevent_openssl.c: In function 'bio_bufferevent_new':
    bufferevent_openssl.c:106:3: error: dereferencing pointer to incomplete type
    b->init = 0;

libevent-2.0.21-5.xs8.src.rpm is disabling ssl at buildtime (in configure step),
this is not desired for XCP-ng as explained above.

To minimize potential future divergence with upstream, I made openssl
conditional, and part of this change is under review at upstream.

Origin: https://github.com/xcp-ng-rpms/libevent/pull/2
Relate-to: https://src.fedoraproject.org/rpms/libevent/pull-request/9
Relate-to: https://github.com/xcp-ng-rpms/libevent/pull/1#discussion_r2313257313
Relate-to: XCPNG-2874
Signed-off-by: Philippe Coval <philippe.coval@vates.tech>